\name{EDproject-package}
\alias{EDproject-package}
\alias{EDproject}
\docType{package}
\title{
\packageTitle{EDproject}
}
\description{
\packageDescription{EDproject}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EDproject}
\packageIndices{EDproject}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EDproject}

Maintainer: \packageMaintainer{EDproject}
}
